feat: rework state in app
